requirements.txt
psycopg2-binary>=2.9.0
pymysql>=1.0.2
pyodbc>=4.0.35
sqlalchemy>=1.4.0
lxml>=4.6.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
flask>=2.0.0
numpy>=1.21.0
openpyxl>=3.0.0